+++ CodeStarterCamp_Week1/Step3.swift
@@ -18,7 +18,7 @@ var peperoBody : String = ""
 // MARK: 빼빼로 그리기 요구사항1
 
 // 요구사항 1-1 : 길이에 따른 막대(빼빼로 손잡이)를 그리는 함수
-func printPeperoBar(barLength:Int){
+func printPeperoBar(barLength:Int) {
     for i in 1...barLength{
         print(peperoBar)
     }
@@ -26,25 +26,23 @@ func printPeperoBar(barLength:Int){
 
 
 // 요구사항 1-2 : 몸통(String), 토핑(String)을 매개변수로 받아 토핑과 몸통을 그리는 함수
-func makePeperoBody(body:String, topping:String?) -> String{
+func makePeperoBody(body:String, topping:String?) -> String {
     
+    let lineBreak = "\n "
     if (topping == nil && body.hasPrefix("*")){
-        return " " + body
+        return " " + body + lineBreak + body
     }else if (topping == nil){
-        return peperoBar + "\n " + body
+        return " | |" + lineBreak + body
     }
     
-    let tempBody = topping! + body + "\n " + body + topping!
+    let tempBody = topping! + body + lineBreak + body + topping!
     return tempBody
 }
 
 // 요구사항 1-3 : 길이(Int)에 따라 몸통을 그리는 함수
-func printPeperoBody(bodyLength:Int){
+func printPeperoBody(bodyLength:Int) {
     
-    var tempLength = bodyLength
-    if (!peperoBody.trimmingCharacters(in: .whitespaces).hasPrefix("*")){
-        tempLength /= 2
-    }
+    var tempLength = bodyLength / 2
     
     for _ in 1...tempLength {
         print(peperoBody)
@@ -52,7 +50,7 @@ func printPeperoBody(bodyLength:Int){
 }
 
 // 요구사항 1-4 : 각종 매개변수를 받아 위의 세 메서드를 호출하는 함수
-func printPepero(type:PeperoType){
+func printPepero(type:PeperoType) {
     
     var body: String = "***"
     var topping: String?
